+++ src/youwol/backends/tree_db/root_paths.py
@@ -1,7 +1,6 @@
 # standard library
 import asyncio
 import itertools
-import json
 import uuid
 
 from collections.abc import Coroutine
@@ -16,18 +15,6 @@ from youwol.backends.tree_db.configurations import (
     Constants,
     get_configuration,
 )
-from youwol.backends.tree_db.utils import (
-    db_delete,
-    db_get,
-    db_post,
-    db_query,
-    doc_to_drive_response,
-    doc_to_folder,
-    doc_to_item,
-    folder_to_doc,
-    get_parent,
-    item_to_doc,
-)
 
 # Youwol utilities
 from youwol.utils import (
@@ -39,7 +26,6 @@ from youwol.utils import (
 )
 from youwol.utils.context import Context
 from youwol.utils.http_clients.tree_db_backend import (
-    BorrowBody,
     ChildrenResponse,
     DefaultDriveResponse,
     DriveBody,
@@ -49,20 +35,34 @@ from youwol.utils.http_clients.tree_db_backend import (
     FolderResponse,
     Group,
     GroupsResponse,
-    ItemBody,
     ItemResponse,
-    ItemsResponse,
     PathResponse,
     PurgeResponse,
     RenameBody,
 )
 
 # relative
-from .routers import router_entities
-from .utils import entities_children
+from .routers import router_entities, router_items
+from .utils import (
+    db_delete,
+    db_get,
+    db_post,
+    db_query,
+    doc_to_drive_response,
+    doc_to_folder,
+    entities_children,
+    folder_to_doc,
+    get_drive,
+    get_folder,
+    get_folders_rec,
+    get_parent,
+    item_to_doc,
+    list_deleted,
+)
 
 router = APIRouter(tags=["treedb-backend"])
 router.include_router(router_entities)
+router.include_router(router_items)
 flatten = itertools.chain.from_iterable
 
 
@@ -225,21 +225,6 @@ async def update_drive(
         return response
 
 
-async def _get_drive(drive_id: str, configuration: Configuration, context: Context):
-    async with context.start(action="_get_drive") as ctx:
-        docdb = configuration.doc_dbs.drives_db
-        doc = await db_get(
-            docdb=docdb, partition_keys={"drive_id": drive_id}, context=ctx
-        )
-
-        return DriveResponse(
-            driveId=drive_id,
-            name=doc["name"],
-            metadata=doc["metadata"],
-            groupId=doc["group_id"],
-        )
-
-
 @router.get(
     "/default-drive",
     response_model=DefaultDriveResponse,
@@ -269,7 +254,7 @@ async def get_default_user_drive(
     summary="Retrieves a drive properties.",
     response_model=DriveResponse,
 )
-async def get_drive(
+async def get_drive_details(
     request: Request,
     drive_id: str,
     configuration: Configuration = Depends(get_configuration),
@@ -291,7 +276,7 @@ async def get_drive(
         action="get_drive",
         with_attributes={"drive_id": drive_id},
     ) as ctx:  # type: Context
-        response = await _get_drive(
+        response = await get_drive(
             drive_id=drive_id, configuration=configuration, context=ctx
         )
         return response
@@ -308,7 +293,7 @@ async def _ensure_folder(
         action="ensure folder", with_attributes={"folder_id": folder_id, "name": name}
     ) as ctx:
         try:
-            folder_resp = await _get_folder(
+            folder_resp = await get_folder(
                 folder_id=folder_id, configuration=configuration, context=context
             )
             await ctx.info("Folder already exists")
@@ -337,7 +322,7 @@ async def _get_default_drive(
     ) as ctx:
         default_drive_id = f"{group_id}_default-drive"
         try:
-            await _get_drive(
+            await get_drive(
                 drive_id=default_drive_id, configuration=configuration, context=ctx
             )
         except HTTPException as e_drive:
@@ -541,22 +526,12 @@ async def update_folder(
         return response
 
 
-async def _get_folder(folder_id: str, configuration: Configuration, context: Context):
-    async with context.start(action="_get_folder") as ctx:
-        doc = await db_get(
-            partition_keys={"folder_id": folder_id},
-            context=ctx,
-            docdb=configuration.doc_dbs.folders_db,
-        )
-        return doc_to_folder(doc)
-
-
 @router.get(
     "/folders/{folder_id}",
     summary="Retrieves properties of a folder.",
     response_model=FolderResponse,
 )
-async def get_folder(
+async def get_folder_details(
     request: Request,
     folder_id: str,
     configuration: Configuration = Depends(get_configuration),
@@ -577,250 +552,9 @@ async def get_folder(
         action="get_folder",
         with_attributes={"folder_id": folder_id},
     ) as ctx:  # type: Context
-        response = await _get_folder(
+        response = await get_folder(
             folder_id=folder_id, configuration=configuration, context=ctx
         )
-        return response
-
-
-async def _create_item(
-    folder_id: str, item: ItemBody, configuration: Configuration, context: Context
-):
-    async with context.start(action="_create_item") as ctx:
-        items_db = configuration.doc_dbs.items_db
-        parent = await get_parent(
-            parent_id=folder_id, configuration=configuration, context=ctx
-        )
-
-        doc = {
-            "item_id": item.itemId or str(uuid.uuid4()),
-            "folder_id": folder_id,
-            "related_id": item.assetId,
-            "name": item.name,
-            "type": item.kind,
-            "group_id": parent["group_id"],
-            "drive_id": parent["drive_id"],
-            "metadata": json.dumps({"borrowed": item.borrowed}),
-        }
-        await db_post(docdb=items_db, doc=doc, context=ctx)
-
-        response = doc_to_item(doc)
-        return response
-
-
-@router.put(
-    "/folders/{folder_id}/items",
-    summary="Create a new item.",
-    response_model=ItemResponse,
-)
-async def create_item(
-    request: Request,
-    folder_id: str,
-    item: ItemBody,
-    configuration: Configuration = Depends(get_configuration),
-) -> ItemResponse:
-    """
-    Creates a new item.
-
-    Note:
-        An item is related to an asset: before creating one, the corresponding asset should have been created
-            (and its ID provided in the `item` parameter).
-
-    Parameters:
-        request: Incoming request.
-        folder_id: ID of the parent folder.
-        item: item properties.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the folder.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="create_item",
-        body=item,
-        with_attributes={"folder_id": folder_id},
-    ) as ctx:  # type: Context
-        return await _create_item(
-            folder_id=folder_id, item=item, configuration=configuration, context=ctx
-        )
-
-
-@router.post(
-    "/items/{item_id}",
-    summary="Updates an item properties.",
-    response_model=ItemResponse,
-)
-async def update_item(
-    request: Request,
-    item_id: str,
-    body: RenameBody,
-    configuration: Configuration = Depends(get_configuration),
-) -> ItemResponse:
-    """
-    Updates an item properties.
-
-    Parameters:
-        request: Incoming request.
-        item_id: ID of the item.
-        body: Update details.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the item.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="update_item",
-        body=body,
-        with_attributes={"item_id": item_id},
-    ) as ctx:  # type: Context
-        items_db = configuration.doc_dbs.items_db
-        doc = await db_get(
-            partition_keys={"item_id": item_id}, docdb=items_db, context=ctx
-        )
-        doc = {**doc, **{"name": body.name}}
-        await db_post(docdb=items_db, doc=doc, context=ctx)
-
-        response = doc_to_item(doc)
-        return response
-
-
-async def _get_item(item_id: str, configuration: Configuration, context: Context):
-    async with context.start(action="_get_item") as ctx:  # type: Context
-        items_db = configuration.doc_dbs.items_db
-        doc = await db_get(
-            partition_keys={"item_id": item_id}, docdb=items_db, context=ctx
-        )
-        return doc_to_item(doc)
-
-
-@router.get(
-    "/items/{item_id}",
-    summary="Retrieves properties of an item",
-    response_model=ItemResponse,
-)
-async def get_item(
-    request: Request,
-    item_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> ItemResponse:
-    """
-    Retrieves properties of an item.
-
-    Parameters:
-        request: Incoming request.
-        item_id: ID of the item.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the item.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="get_item",
-        with_attributes={"item_id": item_id},
-    ) as ctx:  # type: Context
-        response = await _get_item(
-            item_id=item_id, configuration=configuration, context=ctx
-        )
-        return response
-
-
-@router.get(
-    "/items/from-asset/{asset_id}", summary="get an item", response_model=ItemsResponse
-)
-async def get_items_by_asset_id(
-    request: Request,
-    asset_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> ItemsResponse:
-    """
-    Retrieves the list of items associated to a corresponding `assetID`.
-    From this list, one is the original item (not borrowed), the others are symbolic links (borrowed).
-
-    Parameters:
-        request: Incoming request.
-        asset_id: ID of the corresponding asset.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the item.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="get_item",
-        with_attributes={"assetId": asset_id},
-    ) as ctx:  # type: Context
-        docdb = configuration.doc_dbs.items_db
-        items = await db_query(
-            docdb=docdb, key="related_id", value=asset_id, max_count=100, context=ctx
-        )
-
-        response = ItemsResponse(items=[doc_to_item(item) for item in items])
-        return response
-
-
-async def get_folders_rec(
-    folder_id: str, drive_id: str, configuration: Configuration, context: Context
-):
-    drive = await _get_drive(
-        drive_id=drive_id, configuration=configuration, context=context
-    )
-
-    folders = [
-        await _get_folder(
-            folder_id=folder_id, configuration=configuration, context=context
-        )
-    ]
-    while folders[0].parentFolderId != folders[0].driveId:
-        folders = [
-            await _get_folder(
-                folder_id=folders[0].parentFolderId,
-                configuration=configuration,
-                context=context,
-            )
-        ] + folders
-    return folders, drive
-
-
-@router.get(
-    "/items/{item_id}/path",
-    summary="get the path of an item",
-    response_model=PathResponse,
-)
-async def get_path(
-    request: Request,
-    item_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> PathResponse:
-    """
-    Retrieves the full path of an item.
-
-    Parameters:
-        request: Incoming request.
-        item_id: ID of the item.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the path.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="get_path",
-        with_attributes={"item_id": item_id},
-    ) as ctx:  # type: Context
-        item = await _get_item(
-            item_id=item_id, configuration=configuration, context=ctx
-        )
-        folders, drive = await get_folders_rec(
-            folder_id=item.folderId,
-            drive_id=item.driveId,
-            configuration=configuration,
-            context=ctx,
-        )
-
-        response = PathResponse(item=item, folders=folders, drive=drive)
         return response
 
 
@@ -850,7 +584,7 @@ async def get_path_folder(
         action="get_path_folder",
         with_attributes={"folder_id": folder_id},
     ) as ctx:  # type: Context
-        folder = await _get_folder(
+        folder = await get_folder(
             folder_id=folder_id, configuration=configuration, context=ctx
         )
         folders, drive = await get_folders_rec(
@@ -862,50 +596,6 @@ async def get_path_folder(
 
         response = PathResponse(folders=folders, drive=drive)
         return response
-
-
-@router.post(
-    "/items/{item_id}/borrow", response_model=ItemResponse, summary="borrow item"
-)
-async def borrow(
-    request: Request,
-    item_id: str,
-    body: BorrowBody,
-    configuration: Configuration = Depends(get_configuration),
-) -> ItemResponse:
-    """
-    Create a symbolic link of an item.
-
-    Parameters:
-        request: Incoming request.
-        item_id: Item's ID.
-        body: Borrow specification
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the resulting item.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="borrow item",
-        body=body,
-        with_attributes={"item_id": item_id},
-    ) as ctx:
-        item = await _get_item(
-            item_id=item_id, configuration=configuration, context=ctx
-        )
-
-        metadata = json.loads(item.metadata)
-        metadata["borrowed"] = True
-        item.itemId = body.targetId if body.targetId else str(uuid.uuid4())
-        item.borrowed = True
-        item.metadata = json.dumps(metadata)
-        return await _create_item(
-            folder_id=body.destinationFolderId,
-            item=ItemBody(**item.dict()),
-            configuration=configuration,
-            context=ctx,
-        )
 
 
 @router.get(
@@ -936,123 +626,6 @@ async def children(
             folder_id=folder_id, configuration=configuration, context=ctx
         )
         return response
-
-
-async def _list_deleted(drive_id: str, configuration: Configuration, context: Context):
-    async with context.start("_list_deleted") as ctx:  # type: Context
-        doc_dbs = configuration.doc_dbs
-        entities = await doc_dbs.deleted_db.query(
-            query_body=f"drive_id={drive_id}#100",
-            owner=Constants.public_owner,
-            headers=ctx.headers(),
-        )
-
-        folders = [
-            doc_to_folder(f, {"folder_id": f["deleted_id"]})
-            for f in entities["documents"]
-            if f["kind"] == "folder"
-        ]
-
-        items = [
-            doc_to_item(
-                f, {"item_id": f["deleted_id"], "folder_id": f["parent_folder_id"]}
-            )
-            for f in entities["documents"]
-            if f["kind"] == "item"
-        ]
-
-        response = ChildrenResponse(folders=folders, items=items)
-        return response
-
-
-@router.get(
-    "/drives/{drive_id}/deleted",
-    summary="list items of the drive queued for deletion",
-    response_model=ChildrenResponse,
-)
-async def list_deleted(
-    request: Request,
-    drive_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> ChildrenResponse:
-    """
-    Query the entities queued for deletion (moved into the 'trash').
-
-    Parameters:
-        request: Incoming request
-        drive_id: parent drive's ID of the 'trash'.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the children.
-    """
-    async with Context.start_ep(
-        request=request, action="list_deleted"
-    ) as ctx:  # type: Context
-        response = await _list_deleted(
-            drive_id=drive_id, configuration=configuration, context=ctx
-        )
-        return response
-
-
-@router.delete(
-    "/items/{item_id}",
-    summary="Queues an entity for deletion (moves into the 'trash').",
-)
-async def queue_delete_item(
-    request: Request,
-    item_id: str,
-    erase: bool = False,
-    configuration: Configuration = Depends(get_configuration),
-):
-    """
-    Queues an item for deletion (moves into the 'trash').
-
-    Parameters:
-        request: Incoming request
-        item_id: Item's ID.
-        erase: if `True`, the entity is deleted directly (and not queued for deletion).
-        configuration: Injected configuration of the service.
-
-    Return:
-        Empty JSON response.
-    """
-    async with Context.start_ep(
-        request=request, action="queue_delete_item", with_attributes={"itemId": item_id}
-    ) as ctx:  # type: Context
-        dbs = configuration.doc_dbs
-        items_db, _, _, deleted_db = (
-            dbs.items_db,
-            dbs.folders_db,
-            dbs.drives_db,
-            dbs.deleted_db,
-        )
-
-        doc = await db_get(
-            partition_keys={"item_id": item_id}, docdb=items_db, context=ctx
-        )
-
-        if not erase:
-            deleted_doc = {
-                "deleted_id": doc["item_id"],
-                "drive_id": doc["drive_id"],
-                "type": doc["type"],
-                "kind": "item",
-                "related_id": doc["related_id"],
-                "name": doc["name"],
-                "parent_folder_id": doc["folder_id"],
-                "group_id": doc["group_id"],
-                "metadata": doc["metadata"],
-            }
-            deleted_db = configuration.doc_dbs.deleted_db
-            await db_post(doc=deleted_doc, docdb=deleted_db, context=ctx)
-
-        await db_delete(
-            docdb=items_db,
-            doc={"item_id": doc["item_id"], "group_id": doc["group_id"]},
-            context=ctx,
-        )
-        return {}
 
 
 @router.delete("/folders/{folder_id}", summary="delete a folder and its content")
@@ -1139,7 +712,7 @@ async def delete_drive(
             entities_children(
                 folder_id=drive_id, configuration=configuration, context=ctx
             ),
-            _list_deleted(drive_id=drive_id, configuration=configuration, context=ctx),
+            list_deleted(drive_id=drive_id, configuration=configuration, context=ctx),
         )
 
         if len(entities.folders + entities.items + deleted.items + deleted.folders) > 0:
@@ -1184,7 +757,7 @@ async def purge_drive(
         dbs = configuration.doc_dbs
         folders_db, items_db = dbs.folders_db, dbs.items_db
 
-        deleted = await _list_deleted(
+        deleted = await list_deleted(
             drive_id=drive_id, configuration=configuration, context=ctx
         )
