+++ src/youwol/backends/tree_db/routers/__init__.py
@@ -1,2 +1,3 @@
 # relative
 from .entities import router as router_entities
+from .items import router as router_items
